Refactor ThemesController: Ekstraksi logika bisnis dan validasi ke layanan terpisah
Commit ini memindahkan logika bisnis manajemen tema ke ThemeService, validasi keamanan ke ThemeHooksValidator, dan menggunakan CacheService untuk pengelolaan cache agar controller lebih bersih dan terstruktur.
